Split position sensor
